Проведён рефакторинг архитектуры MVVM. ВидМодели для сообщений и диалогов объединены в один MainViewModel для основной формы.
