+++ VKMessenger/VKMessenger.csproj
@@ -96,8 +96,6 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="Model\Dialogs.cs" />
-    <Compile Include="Model\Messages.cs" />
     <Compile Include="Model\VkMessage.cs" />
     <Compile Include="Protocol\DVProto.cs" />
     <Compile Include="Protocol\IEndToEndProtocol.cs" />
@@ -110,9 +108,7 @@
     <Compile Include="ViewModel\Commands\SimpleCommand.cs" />
     <Compile Include="ViewModel\Design\DialogsDesignViewModel.cs" />
     <Compile Include="ViewModel\Design\MessagesDesignViewModel.cs" />
-    <Compile Include="ViewModel\DialogsLoader.cs" />
     <Compile Include="ViewModel\MainViewModel.cs" />
-    <Compile Include="ViewModel\MessagesLoader.cs" />
     <Compile Include="ViewModel\SettingsViewModel.cs" />
     <Compile Include="View\AuthorizationWindow.xaml.cs">
       <DependentUpon>AuthorizationWindow.xaml</DependentUpon>
