+++ VKMessenger/ViewModel/MainViewModel.cs
@@ -1,9 +1,15 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
+using System.Threading.Tasks;
 using VKMessenger.Model;
 using VKMessenger.ViewModel.Commands;
 using VkNet;
+using VkNet.Enums.Filters;
+using VkNet.Model;
+using VkNet.Model.RequestParams;
 
 namespace VKMessenger.ViewModel
 {
@@ -19,6 +25,9 @@ namespace VKMessenger.ViewModel
         }
     }
 
+	/// <summary>
+	/// Бизнес-логика основного окна мессенджера.
+	/// </summary>
 	public class MainViewModel : INotifyPropertyChanged
     {
         private Messenger _messenger;
@@ -32,43 +41,13 @@ namespace VKMessenger.ViewModel
 					_messenger = value;
 					_messenger.NewMessage += ReceiveNewMessage;
 					_messenger.MessageSent += NewMessageSent;
-					DialogsViewModel.Messenger = _messenger;
-					MessagesViewModel.Messenger = _messenger;
-					DialogsViewModel.LoadDialogs();
+					LoadDialogs();
 					OnPropertyChanged();
 				}
 			}
 		}
 
 		public VkApi Vk { get { return _messenger.Vk; } }
-
-        private DialogsLoader _dialogsViewModel = new DialogsLoader();
-        public DialogsLoader DialogsViewModel
-        {
-            get { return _dialogsViewModel; }
-            set
-            {
-                if (_dialogsViewModel != value)
-                {
-                    _dialogsViewModel = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
-
-        private MessagesLoader _messagesViewModel = new MessagesLoader();
-        public MessagesLoader MessagesViewModel
-        {
-            get { return _messagesViewModel; }
-            set
-            {
-                if (_messagesViewModel != value)
-                {
-                    _messagesViewModel = value;
-                    OnPropertyChanged();
-                }
-            }
-        }
 
         private string _messageText;
 		/// <summary>
@@ -87,32 +66,100 @@ namespace VKMessenger.ViewModel
                 }
             }
         }
-		
+
+		private int _selectedDialogIndex = -1;
+		/// <summary>
+		/// Индекс выбранного диалога в списке.
+		/// </summary>
+		public int SelectedDialogIndex
+		{
+			get { return _selectedDialogIndex; }
+			set
+			{
+				if (_selectedDialogIndex != value)
+				{
+					_selectedDialogIndex = value;
+					OnDialogChanged();
+					OnPropertyChanged();
+				}
+			}
+		}
+
+		private ObservableCollection<Dialog> _dialogs = new ObservableCollection<Dialog>();
+		/// <summary>
+		/// Список диалогов пользователя.
+		/// </summary>
+		public ObservableCollection<Dialog> Dialogs
+		{
+			get { return _dialogs; }
+			set
+			{
+				if (_dialogs != value)
+				{
+					_dialogs = value;
+					OnPropertyChanged();
+				}
+			}
+		}
+
+		/// <summary>
+		/// Выбранный диалог.
+		/// </summary>
+		public Dialog SelectedDialog
+		{
+			get { return SelectedDialogIndex >= 0 ? Dialogs[SelectedDialogIndex] : null; }
+		}
+
+		private ObservableCollection<VkMessage> _messages = new ObservableCollection<VkMessage>();
+		/// <summary>
+		/// Список сообщений выбранного диалога.
+		/// </summary>
+		public ObservableCollection<VkMessage> Messages
+		{
+			get { return _messages; }
+			set
+			{
+				if (_messages != value)
+				{
+					_messages = value;
+					OnPropertyChanged();
+				}
+			}
+		}
+
+		#region Команды
 		/// <summary>
 		/// Команда отправки нового сообщения.
 		/// </summary>
-        public SimpleCommand SendMessageCommand { get; set; }
+		public SimpleCommand SendMessageCommand { get; set; }
+		#endregion
 
+		#region События
 		/// <summary>
 		/// Вызывается при получении нового сообщения.
 		/// </summary>
 		public event EventHandler<NewMessageEventArgs> NewMessage;
+		#endregion
 
-        public event PropertyChangedEventHandler PropertyChanged;
+		public event PropertyChangedEventHandler PropertyChanged;
 
         public MainViewModel()
         {
-            DialogsViewModel.PropertyChanged += DialogsViewModel_PropertyChanged;
             SendMessageCommand = new SimpleCommand(SendMessageExecute, CanSendMessage);
         }
+
+		protected virtual void OnPropertyChanged([CallerMemberName]string propertyName = "")
+		{
+			PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+		}
 
 		/// <summary>
 		/// Обрабатывает новое сообщение.
 		/// </summary>
-        private void ReceiveNewMessage(object sender, MessageEventArgs e)
+		private void ReceiveNewMessage(object sender, MessageEventArgs e)
         {
             VkMessage message = e.Message;
-            Dialog currentDialog = DialogsViewModel.SelectedDialog;
+            Dialog currentDialog = SelectedDialog;
 
             if (currentDialog != null
                 && ((currentDialog.IsChat && currentDialog.Chat.Id == message.Content.ChatId)
@@ -122,13 +169,13 @@ namespace VKMessenger.ViewModel
 
                 System.Windows.Application.Current.Dispatcher.Invoke(() =>
                 {
-                    currentDialog.Messages.Content.Add(message);
+                    currentDialog.Messages.Add(message);
                 });
             }
 			
             Dialog dialogForMessage = currentDialog;
 
-			foreach (Dialog dialog in DialogsViewModel.Model.Content)
+			foreach (Dialog dialog in Dialogs)
 			{
 				if ((dialog.IsChat && dialog.Chat.Id == message.Content.ChatId)
 					|| (!dialog.IsChat && dialog.User.Id == message.Content.UserId.Value))
@@ -151,22 +198,6 @@ namespace VKMessenger.ViewModel
 			ReceiveNewMessage(sender, e);
 		}
 
-        private void DialogsViewModel_PropertyChanged(object sender, PropertyChangedEventArgs e)
-        {
-            switch (e.PropertyName)
-            {
-                case nameof(DialogsViewModel.SelectedDialogIndex):
-                    MessagesViewModel.Dialog = DialogsViewModel.SelectedDialog;
-                    SendMessageCommand.RaiseCanExecuteChanged();
-                    break;
-            }
-        }
-
-        protected virtual void OnPropertyChanged([CallerMemberName]string propertyName = "")
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        }
-
 		/// <summary>
 		/// Вызывает событие получения нового сообщения.
 		/// </summary>
@@ -182,7 +213,7 @@ namespace VKMessenger.ViewModel
 		/// </summary>
         private async void SendMessageExecute()
         {
-			await _messenger.SendMessage(MessageText, DialogsViewModel.SelectedDialog);
+			await _messenger.SendMessage(MessageText, SelectedDialog);
 			MessageText = string.Empty;
 		}
 
@@ -192,7 +223,156 @@ namespace VKMessenger.ViewModel
 		/// <returns></returns>
         private bool CanSendMessage()
         {
-            return !string.IsNullOrWhiteSpace(MessageText) && DialogsViewModel.SelectedDialog != null;
+            return !string.IsNullOrWhiteSpace(MessageText) && SelectedDialog != null;
         }
-    }
+
+		private async void LoadDialogs()
+		{
+			ReadOnlyCollection<Message> dialogMessages = await GetDialogsList();
+
+			for (int i = 0; i < dialogMessages.Count; i++)
+			{
+				Message lastMessage = dialogMessages[i];
+
+				if (lastMessage.UserId < 0)
+				{
+					continue;
+				}
+
+				Dialog dialog = new Dialog();
+				VkMessage message = new VkMessage(lastMessage, dialog);
+
+				if (lastMessage.ChatId.HasValue)
+				{
+					dialog.Photo = lastMessage.Photo50;
+					dialog.Chat = await LoadChat(lastMessage.ChatId.Value);
+					dialog.Users = await LoadUsers(dialog.Chat.Users);
+				}
+				else
+				{
+					dialog.User = await LoadUser(lastMessage.UserId.Value);
+				}
+
+
+				Dialogs.Add(dialog);
+			}
+		}
+
+		private Task<List<User>> LoadUsers(Collection<long> usersIds)
+		{
+			return Task.Run(() =>
+			{
+				List<User> users = new List<User>(usersIds.Count);
+
+				Utils.Extensions.BeginVkInvoke(Vk);
+				ReadOnlyCollection<User> usersCollection = Vk.Users.Get(usersIds, ProfileFields.FirstName | ProfileFields.LastName | ProfileFields.Photo50);
+				Utils.Extensions.EndVkInvoke();
+
+				users.AddRange(usersCollection);
+
+				return users;
+			});
+		}
+
+		private Task<User> LoadUser(long userId)
+		{
+			return Task.Run(() =>
+			{
+				Utils.Extensions.BeginVkInvoke(Vk);
+				User user = Vk.Users.Get(userId, ProfileFields.FirstName | ProfileFields.LastName | ProfileFields.Photo50);
+				Utils.Extensions.EndVkInvoke();
+
+				return user;
+			});
+		}
+
+		private Task<Chat> LoadChat(long chatId)
+		{
+			return Task.Run(() =>
+			{
+				Utils.Extensions.BeginVkInvoke(Vk);
+				Chat chat = Vk.Messages.GetChat(chatId);
+				Utils.Extensions.EndVkInvoke();
+
+				return chat;
+			});
+		}
+
+		private Task<ReadOnlyCollection<Message>> GetDialogsList()
+		{
+			return Task.Run(() =>
+			{
+				List<Message> messages = new List<Message>();
+				int offset = 0;
+
+				Utils.Extensions.BeginVkInvoke(Vk);
+				MessagesGetObject response = Vk.Messages.GetDialogs(new MessagesDialogsGetParams()
+				{
+					PreviewLength = 1,
+					Count = 20
+				});
+				Utils.Extensions.EndVkInvoke();
+
+				offset += response.Messages.Count;
+
+				messages.AddRange(response.Messages);
+
+				while (offset + response.Messages.Count < response.TotalCount)
+				{
+					Utils.Extensions.BeginVkInvoke(Vk);
+					response = Vk.Messages.GetDialogs(new MessagesDialogsGetParams()
+					{
+						Offset = offset,
+						PreviewLength = 1,
+						Count = 20
+					});
+					Utils.Extensions.EndVkInvoke();
+
+					offset += response.Messages.Count;
+
+					messages.AddRange(response.Messages);
+				}
+
+				return new ReadOnlyCollection<Message>(messages);
+			});
+		}
+
+		public void SetData(IEnumerable<Message> messages, Dialog dialog)
+		{
+			_messages.Clear();
+
+			foreach (Message message in messages)
+			{
+				_messages.Insert(0, new VkMessage(message, dialog));
+			}
+		}
+
+		protected virtual async void OnDialogChanged()
+		{
+			long peerId = SelectedDialog.PeerId;
+			List<Message> messages = await LoadHistory(peerId);
+
+			SetData(messages, SelectedDialog);
+			SelectedDialog.Messages = Messages;
+		}
+
+		private Task<List<Message>> LoadHistory(long peerId)
+		{
+			return Task.Run(() =>
+			{
+				List<Message> messages = new List<Message>();
+
+				Utils.Extensions.BeginVkInvoke(Vk);
+				MessagesGetObject history = Vk.Messages.GetHistory(new MessagesGetHistoryParams()
+				{
+					PeerId = peerId,
+					Count = 10
+				});
+				Utils.Extensions.EndVkInvoke();
+				messages.AddRange(history.Messages);
+
+				return messages;
+			});
+		}
+	}
 }
